+++ src/strongbox/ui/cli.clj
@@ -605,6 +605,10 @@
 (def pretty-dt-printer (doto (PrettyTime.)
                          (.removeUnit Decade)))
 
+(defn format-dt
+  [val]
+  (some->> val utils/nilable utils/todt (.format pretty-dt-printer)))
+
 (def column-map
   {:browse-local {:label "browse" :value-fn :addon-dir}
    :source {:label "source" :value-fn :source}
@@ -614,8 +618,8 @@
    :tag-list {:label "tags" :value-fn (fn [row]
                                         (when-not (empty? (:tag-list row))
                                           (str (:tag-list row))))}
-   :updated-date {:label "updated" :value-fn #(some->> % :updated-date utils/todt (.format pretty-dt-printer))}
-   :created-date {:label "created" :value-fn #(some->> % :created-date utils/todt (.format pretty-dt-printer))}
+   :updated-date {:label "updated" :value-fn (comp format-dt :updated-date)}
+   :created-date {:label "created" :value-fn (comp format-dt :created-date)}
    :installed-version {:label "installed" :value-fn :installed-version}
    :available-version {:label "available" :value-fn available-versions-v1}
    :combined-version {:label "version" :value-fn available-versions-v2}